Modernize the plugin, mostly packaging with some small extras
- `setup.cfg` + `setup.py` replaced with `pyproject.toml` metadata
- Added installation & usage sections to the README file (which also
  becomes the PyPI package description)
- Moved the plugin from `sopel_pypi` to `sopel_pypi.plugin`, as is the
  modern convention in our cookiecutter template (`gh:sopel-irc/plugin`)
- Updated the plugin file header with copyright & license info
- Removed Python 2-style things (`#coding` magic comment, obsolete
  `__future__` imports)
- Removed the `match` parameter from URL handler; those use the regular
  `trigger` groups now in Sopel 8
- Declared minimum support for Sopel 8.0+ and Python 3.8+
- Actually declared `requests` as a direct plugin dependency, where it
  was previously just assumed to be available because Sopel has had it
  as a dependency for several years
