+++ pyproject.toml
@@ -1,0 +1,51 @@
+[build-system]
+requires = ["setuptools>=63.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+platforms = ["Linux x86, x86-64"]
+
+[tool.setuptools.packages.find]
+include = ["sopel_pypi", "sopel_pypi.*"]
+namespaces = false
+
+[tool.setuptools.dynamic]
+readme = { file=["README.md", "NEWS"], content-type="text/markdown" }
+
+[project]
+name = "sopel-pypi"
+version = "0.2.0"
+description = "A Sopel plugin to show information about linked PyPI packages."
+
+authors = [
+  { name="dgw", email="dgw@technobabbl.es" },
+]
+
+license = "EFL-2.0"
+license-files = ["COPYING"]
+dynamic = ["readme"]
+
+classifiers = [
+  "Intended Audience :: Developers",
+  "Intended Audience :: System Administrators",
+  "Topic :: Communications :: Chat :: Internet Relay Chat",
+]
+keywords = [
+  "sopel",
+  "plugin",
+  "bot",
+  "irc",
+]
+
+requires-python = ">=3.8, <4"
+dependencies = [
+  "sopel>=8.0",
+  "requests",  # let Sopel's requirements determine the version
+]
+
+[project.urls]
+"Homepage" = "https://github.com/sopel-irc/sopel-pypi"
+"Bug Tracker" = "https://github.com/sopel-irc/sopel-pypi/issues"
+
+[project.entry-points."sopel.plugins"]
+"pypi" = "sopel_pypi.plugin"
