--- setup.cfg
@@ -1,25 +1,0 @@
-[metadata]
-name = sopel-pypi
-version = 0.2.0
-description = A Sopel plugin to show information about linked PyPI packages.
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/sopel-irc/sopel-pypi
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.1,<9
-
-[options.entry_points]
-sopel.plugins =
-    pypi = sopel_pypi
